# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Documentação da API Laravel para integração com frontend",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria uma nova conta de usuário e autentica automaticamente",
                "operationId": "register",
                "requestBody": {
                    "description": "Dados do novo usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuário",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "description": "E-mail do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "joao.silva@example.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário (mínimo 8 caracteres)",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "SenhaSegura123!"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmação da senha",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "SenhaSegura123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Usuário registrado e autenticado com sucesso",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Cookie de sessão configurado",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password confirmation does not match."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            }
        },
        "/api/verify-email/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Verificar endereço de e-mail",
                "description": "Verifica o endereço de e-mail do usuário usando o link enviado por e-mail",
                "operationId": "verifyEmail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash de verificação",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "a1b2c3d4e5f6g7h8i9j0"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Timestamp de expiração",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1640995200
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "Assinatura da URL",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "E-mail verificado com sucesso - redirecionamento",
                        "headers": {
                            "Location": {
                                "description": "URL de redirecionamento para o frontend",
                                "schema": {
                                    "type": "string",
                                    "example": "http://frontend.app/dashboard?verified=1"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Link inválido ou expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid verification link."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use o token JWT para autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}